+++ CHANGELOG.md
@@ -5,8 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [3.0.14] 2021-09-27
+### Fixed
+- Deeplink url for calendar events [#751](https://github.com/rokwire/illinois-app/issues/751).
+
+## [3.0.14] 2021-09-27
 ### Changed
 - Athletics: do not show "Free admission" when there is no value for "tickets" url [#777](https://github.com/rokwire/illinois-app/issues/777).
+## Added
+- Implemented event detail from DeepLink [#773](https://github.com/rokwire/illinois-app/issues/773).
 
 ## [3.0.13] 2021-09-24
 ## Added